fix: Diary 관련 각종 버그 수정

+++ backend/src/main/java/org/nexters/cultureland/api/service/DiaryService.java
@@ -49,7 +49,7 @@ public class DiaryService {
         return repositoryService.countByUserGroupedMonth(userId, year);
     }
 
-    public HashMap<String, Integer> countByUserGroupedCategory(Long userId) {
+    public HashMap<String, Long> countByUserGroupedCategory(Long userId) {
         return repositoryService.countByUserGroupedCategory(userId);
     }
 }
